feat(Volume): Added the volume cmd
- Added the possibility the you can list, create, attach, detach and delete a volume
- Add a utility, which gives the possibility to request confirmation when deleting a resource
- Modified all cmd to delete, to add confirmation

BREAKING CHANGE: No

+++ cmd/domain_record_remove.go
@@ -21,29 +21,34 @@ var domainRecordRemoveCmd = &cobra.Command{
 			os.Exit(1)
 		}
 
-		domain, err := client.FindDNSDomain(args[0])
-		if err != nil {
-			fmt.Printf("Unable to find domain for your search: %s\n", aurora.Red(err))
-			os.Exit(1)
+		if utility.AskForConfirmDelete("domain record") == nil {
+			domain, err := client.FindDNSDomain(args[0])
+			if err != nil {
+				fmt.Printf("Unable to find domain for your search: %s\n", aurora.Red(err))
+				os.Exit(1)
+			}
+
+			record, err := client.GetDNSRecord(domain.ID, args[1])
+			if err != nil {
+				fmt.Printf("Unable to get domains record: %s\n", aurora.Red(err))
+				os.Exit(1)
+			}
+
+			_, err = client.DeleteDNSRecord(record)
+
+			ow := utility.NewOutputWriterWithMap(map[string]string{"ID": record.ID, "Name": record.Name})
+
+			switch outputFormat {
+			case "json":
+				ow.WriteSingleObjectJSON()
+			case "custom":
+				ow.WriteCustomOutput(outputFields)
+			default:
+				fmt.Printf("The domain record called %s with ID %s was delete\n", aurora.Green(record.Name), aurora.Green(record.ID))
+			}
+		} else {
+			fmt.Println("Operation aborted.")
 		}
 
-		record, err := client.GetDNSRecord(domain.ID, args[1])
-		if err != nil {
-			fmt.Printf("Unable to get domains record: %s\n", aurora.Red(err))
-			os.Exit(1)
-		}
-
-		_, err = client.DeleteDNSRecord(record)
-
-		ow := utility.NewOutputWriterWithMap(map[string]string{"ID": record.ID, "Name": record.Name})
-
-		switch outputFormat {
-		case "json":
-			ow.WriteSingleObjectJSON()
-		case "custom":
-			ow.WriteCustomOutput(outputFields)
-		default:
-			fmt.Printf("The domain record called %s with ID %s was delete\n", aurora.Green(record.Name), aurora.Green(record.ID))
-		}
 	},
 }
